fix(mrmr): argmax NaN guard, aggregator validation, int64 support, dup-col + polars guards (critique batch B)
- EX-1: _is_argmax_eligible tested the pre-filtered finite subset (always True), so NaN columns were never excluded from argmax/gate FE; now tests the whole column.
- S-F5: an unrecognised redundancy_aggregator (e.g. 'JMIM') silently fell through to plain Fleuret; now raises ValueError.
- ST-3: empty-screen fallback support_ now int64 (was int32 on Windows, inconsistent with every other support_ assignment).
- FE-F7: drop a temporal winner name colliding with an existing X column (dedup X_ta) so X stays uniquely labelled.
- EN-2: guard the polars branches of the encoding-recipe extractors on pandas presence (pd None -> clear TypeError, not AttributeError).

+++ src/mlframe/feature_selection/filters/_mrmr_fit_impl/_finalise.py
@@ -278,7 +278,9 @@ def _finalise_empty_support_fallback(self, n_engineered_out, cols, data, nbins, 
                     _raw_mi_all.sort(key=lambda kv: (-kv[1], kv[0]))
                     _topk = [_raw_mi_all[0][0]]
             if _topk:
-                self.support_ = np.array(_topk)
+                # int64 to match every other support_ assignment in the fit body; a bare np.array(list[int]) is
+                # int32 on Windows, an inconsistency that can bite dtype-sensitive downstream concatenation.
+                self.support_ = np.array(_topk, dtype=np.int64)
                 self.n_features_ = len(_topk) + n_engineered_out
                 self.fallback_used_ = True
                 _top_mi = float(_raw_mi[0][1]) if _raw_mi else 0.0
